requirements.txt
google-genai>=1.0.0
pydantic>=2.7
python-dotenv>=1.0
tqdm>=4.66
